change metric to long to accomodate unsigned int max

+++ projects/batfish-common-protocol/src/main/java/org/batfish/datamodel/OspfRoute.java
@@ -14,7 +14,7 @@ public abstract class OspfRoute extends AbstractRoute {
 
   protected final int _admin;
 
-  protected final int _metric;
+  protected final long _metric;
 
   protected final Ip _nextHopIp;
 
@@ -23,7 +23,7 @@ public abstract class OspfRoute extends AbstractRoute {
       @JsonProperty(PROP_NETWORK) Prefix network,
       @JsonProperty(PROP_NEXT_HOP_IP) Ip nextHopIp,
       @JsonProperty(PROP_ADMINISTRATIVE_COST) int admin,
-      @JsonProperty(PROP_METRIC) int metric) {
+      @JsonProperty(PROP_METRIC) long metric) {
     super(network);
     _admin = admin;
     _metric = metric;
@@ -40,7 +40,7 @@ public abstract class OspfRoute extends AbstractRoute {
   @JsonIgnore(false)
   @JsonProperty(PROP_METRIC)
   @Override
-  public final Integer getMetric() {
+  public final Long getMetric() {
     return _metric;
   }
